Component FilterContainer created
